chore: move stx faucet to match btc faucet routes #107
